Removed the requirement for 6 sat fix and reverted back the message to disable all other GPS constellations.
in the ICSPACE2 & 3 flights, I did not observe incredible fluctuation in the number of satellites. However, with the new UBX-CFG-GNSS message, I saw incredible numbers of satellites used to get a fix, up to 23 satellites. I have a feeling that something went wrong with this command and it was possibly using other constellations besides the American GPS satellites to get those fixes. And when it was not getting 23 sat fixes, it was just not getting any fixes at all!. With GPS satellites, at most we can see 12 satellites.